requirements.txt
alembic==1.4.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
dnspython==2.0.0
email-validator==1.1.1
Faker==4.14.0
Flask==1.1.2
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.0
marshmallow-sqlalchemy==0.24.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
urllib3==1.25.11
uuid==1.30
Werkzeug==1.0.1
WTForms==2.3.3



